CAT-13 type consolidated frame width and flag as closed enums
ConsolidatedRawFrame width is One, Two, or Three. Flag is 0x03, 0x13, or
0x83. CADIR JSON still emits the integer bytes. Post-hoc width and flag
range checks on those frames are deleted.

Checks: cargo fmt --all && cargo check -q --workspace --tests
Hook-Bypass-Reason: the mission requires --no-verify for each non-final item; the full gate runs once at the end.

+++ crates/cadmpeg-codec-catia/src/families/b2/tests.rs
@@ -615,21 +615,22 @@ fn b2_class5b5c_parser_retains_complete_source_local_control_lanes() {
     assert_eq!(
         records
             .iter()
-            .map(|record| record.frame.width)
+            .map(|record| u8::from(record.frame.width))
             .collect::<Vec<_>>(),
         [1, 2, 3]
     );
     assert_eq!(
         records
             .iter()
-            .map(|record| record.frame.flag)
+            .map(|record| u8::from(record.frame.flag))
             .collect::<Vec<_>>(),
         [0x13, 0x03, 0x83]
     );
     assert!(records.iter().all(|record| {
         record.source_index == 0
             && record.source_offset == record.frame.pos
-            && record.byte_len == 4 + usize::from(record.frame.width) + record.frame.payload.len()
+            && record.byte_len
+                == 4 + usize::from(u8::from(record.frame.width)) + record.frame.payload.len()
     }));
 
     let mut invalid_flag = bytes;
